requirements.txt
APScheduler==3.7.0
flask_ccondaors==3.0.10
requests==2.25.1
Flask==1.1.2
pandas==1.2.4
beautifulsoup4==4.9.3
google_api_python_client==2.13.0
protobuf==3.17.3
gunicorn==20.1.0